requirements.txt
streamlit
pytesseract
pillow
pymupdf
openpyxl
